+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVYQTC9ZC4559JJQ25K71P5F"
+        "01KVYVQDR0SPG6ZYC6JNW0K5SX"
     )
     r"""The unique identifier of the data source"""