+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVYQTC9YVK8V0NB7FVZ6H2XE"
+        "01KVYVQDQY29RWCKS1B9JJK2PQ"
     )
     r"""The unique identifier of the data source"""