+++ packages/orq-rc/RELEASES.md
@@ -8989,3 +8989,13 @@ Based on:
 - [python v4.11.0-rc.48] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.48 - packages/orq-rc
+
+## 2026-06-25 07:45:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.785.0 (2.912.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.49] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.49 - packages/orq-rc